pyproject.toml
[project]
name = "habit-tracker"
version = "0.1.0"
description = ""
authors = [
    {name = "Oleg Timershayakhov",email = "olegtimer@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "celery (>=5.5.1,<6.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "django (==4.2.2)",
    "django-celery-beat (>=2.7.0,<3.0.0)",
    "ipython (==7.17.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.2.0"
black = "^25.1.0"
isort = "^6.0.1"


[tool.poetry.group.dev.dependencies]
coverage = "^7.8.0"

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | manage.py           # Также отдельно исключить файл в корне проекта
  | poetry.lock          # Также отдельно исключить файл в корне проекта
  | pyproject.toml          # Также отдельно исключить файл в корне проекта
  | .gitignore          # Также отдельно исключить файл в корне проекта

)
'''

[tool.isort]
# максимальная длина строки
line_length = 119
